Move common utilities to OpenLIFULib python module (#55)
