+++ .gitignore
@@ -1,1 +1,5 @@
 build/
+OpenLIFUHome/__pycache__
+OpenLIFULib/OpenLIFULib/__pycache__
+
+
