+++ OpenLIFUHome/CMakeLists.txt
@@ -9,7 +9,6 @@ set(MODULE_PYTHON_SCRIPTS
 set(MODULE_PYTHON_RESOURCES
   Resources/Icons/${MODULE_NAME}.png
   Resources/UI/${MODULE_NAME}.ui
-  Resources/python-requirements.txt
   )
 
 #-----------------------------------------------------------------------------
