+++ OpenLIFULib/OpenLIFULib/Resources/python-requirements.txt
@@ -1,0 +1,1 @@
+git+https://github.com/OpenwaterHealth/OpenLIFU-python.git@38b8deb948fd7107bd8d1ec68896add53e1f1c73
